pyproject.toml
[tool.poetry]
name = "menu-proj"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.3"
psycopg2 = "^2.9.9"
django-autoslug = "^1.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
